// src/types/types.ts
export type AttributeOptionNode = {
  id: string;
  adminName?: string;
  label?: string;
  isValid: boolean;
};

export type AttributeOptionEdge = {
  node: AttributeOptionNode;
};

export type AttributeData = {
  id: string;
  code: string;
  label?: string;
  options: AttributeOptionNode[] | {
    edges: AttributeOptionEdge[];
  };
};